# public/doc/swagger.yaml
swagger: '2.0'
info:
  description: Login Basekit API ドキュメント
  version: 0.0.1
  title: Login Basekit Platform REST APIs
host: localhost:8080
basePath: /
tags:
  - name: user
    description: API for Users
schemes:
  - http
  - https
paths:
  /platformer:
    post:
      tags:
        - user
      summary: 特権ユーザーを新規に作成する
      description: 特権ユーザーを新規に作成する
      parameters:
        - in: body
          name: body
          description: 特権ユーザー情報
          required: true
          schema:
            $ref: '#/definitions/UserRequest'
      responses:
        '200':
          description: 成功
        '400':
          description: エラー
  /login:
    post:
      tags:
        - user
      summary: プラットフォームへのログイン
      description: プラットフォームへのログイン
      parameters:
        - in: body
          name: body
          description: ログイン情報
          required: true
          schema:
            $ref: '#/definitions/LoginRequest'
      responses:
        '200':
          description: 成功
        '400':
          description: エラー
  /logout:
    post:
      tags:
        - user
      summary: プラットフォームからのログアウト
      description: プラットフォームからのログアウト
      parameters:
        - name: x-access-token
          type: string
          in: header
          description: ログイン時に取得したトークン
          required: true
      responses:
        '200':
          description: 成功
        '400':
          description: エラー
  /users:
    get:
      tags:
        - user
      summary: ユーザー情報の取得
      description: ユーザー情報の取得
      parameters:
        - name: x-access-token
          type: string
          in: header
          description: ログイン時に取得したトークン
          required: true
        - name: limit
          type: integer
          in: query
          description: 取得数上限（デフォルトは上限なし）
        - name: offset
          type: integer
          in: query
          description: 取得オフセット（デフォルトは 0）
      responses:
        '200':
          description: 成功
        '400':
          description: エラー
  /user/{id}:
    get:
      tags:
        - user
      summary: 特定のユーザー情報を取り出す
      description: 特定のユーザー情報を取り出す
      parameters:
        - name: x-access-token
          type: string
          in: header
          description: ログイン時に取得したトークン
          required: true
        - in: path
          name: id
          type: string
          description: ユーザーID
          required: true
      responses:
        '200':
          description: 成功
        '400':
          description: エラー
    delete:
      tags:
        - user
      summary: 特定のユーザー情報を削除する
      description: 特定のユーザー情報を削除する
      parameters:
        - name: x-access-token
          type: string
          in: header
          description: ログイン時に取得したトークン
          required: true
        - in: path
          name: id
          type: string
          description: ユーザーID
          required: true
      responses:
        '200':
          description: 成功
        '400':
          description: エラー
  /user:
    post:
      tags:
        - user
      summary: ユーザーを新規に作成する
      description: ユーザーを新規に作成する
      parameters:
        - name: x-access-token
          type: string
          in: header
          description: ログイン時に取得したトークン
          required: true
        - in: body
          name: body
          description: ユーザー情報
          required: true
          schema:
            $ref: '#/definitions/UserRequest'
      responses:
        '200':
          description: 成功
        '400':
          description: エラー
definitions:
  LoginRequest:
    type: object
    properties:
      username:
        type: string
        description: ユーザーID
      password:
        type: string
        description: ユーザーパスワード
  UserRequest:
    type: object
    properties:
      id:
        type: string
        description: ユーザーID
      password:
        type: string
        description: ユーザーパスワード
      role:
        type: integer
        description: ユーザーロール
      name:
        type: string
        description: ユーザー名
      ruby:
        type: string
        description: ユーザー名よみがな
      zip:
        type: string
        description: ユーザー郵便番号
      tel:
        type: string
        description: ユーザー電話番号
      email:
        type: string
        description: ユーザーメールアドレス